+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New   'product-identifier', 'product-identifierLabel', 'product-identifierSeparator', 'product-identifierValue' CSS handles to 'BaseProductIdentifier'
 
 ## [0.1.1] - 2019-08-30